requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.5
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
